[bitmap] Use size_t throughout
This CL changes the interface to always use size_t to be more portable
between 32 and 64 bit platforms.  The only requirement on size_t's size
is in RawBitmap, where it must match unsigned int, unsigned long, or
unsigned long long in order to pick the correct GCC builtin for counting
trailing zeros.

+++ system/ulib/bitmap/include/bitmap/bitmap.h
@@ -4,46 +4,45 @@
 
 #pragma once
 
-#include <stdint.h>
+#include <stddef.h>
 
 #include <magenta/types.h>
-#include <mxtl/macros.h>
 
 namespace bitmap {
 
-// An abstract bitmap..
+// An abstract bitmap.
 class Bitmap {
 public:
     virtual ~Bitmap() = default;
 
     // Returns true in the bit at bitoff is set.
-    virtual bool GetOne(uint64_t bitoff) const {
+    virtual bool GetOne(size_t bitoff) const {
         return Get(bitoff, bitoff + 1, nullptr);
     }
 
     // Returns true if all the bits in [*bitoff*, *bitmax*) are set. Afterwards,
     // *first_unset* will be set to the lesser of bitmax and the index of the
     // first unset bit after *bitoff*.
-    virtual bool Get(uint64_t bitoff, uint64_t bitmax,
-                     uint64_t* first_unset = nullptr) const = 0;
+    virtual bool Get(size_t bitoff, size_t bitmax,
+                     size_t* first_unset = nullptr) const = 0;
 
     // Sets the bit at bitoff.  Only fails on allocation error.
-    virtual mx_status_t SetOne(uint64_t bitoff) {
+    virtual mx_status_t SetOne(size_t bitoff) {
         return Set(bitoff, bitoff + 1);
     }
 
     // Sets all bits in the range [*bitoff*, *bitmax*).  Only fails on
     // allocation error or if bitmax < bitoff.
-    virtual mx_status_t Set(uint64_t bitoff, uint64_t bitmax) = 0;
+    virtual mx_status_t Set(size_t bitoff, size_t bitmax) = 0;
 
     // Clears the bit at bitoff.  Only fails on allocation error.
-    virtual mx_status_t ClearOne(uint64_t bitoff) {
+    virtual mx_status_t ClearOne(size_t bitoff) {
         return Clear(bitoff, bitoff + 1);
     }
 
     // Clears all bits in the range [*bitoff*, *bitmax*).  Only fails on
     // allocation error or if bitmax < bitoff.
-    virtual mx_status_t Clear(uint64_t bitoff, uint64_t bitmax) = 0;
+    virtual mx_status_t Clear(size_t bitoff, size_t bitmax) = 0;
 
     // Clear all bits in the bitmap.
     virtual void ClearAll() = 0;
